move some logic from writeOneTrustAssessment to cli-sync-ot
